+++ src/django_mysql/__init__.py
@@ -1,6 +1,0 @@
-from __future__ import annotations
-
-import django
-
-if django.VERSION < (3, 2):
-    default_app_config = "django_mysql.apps.MySQLConfig"